fixed: win32 warnings.

+++ trunk/src/lib/server/CServer.cpp
@@ -2350,7 +2350,7 @@ CServer::dragInfoReceived(UInt32 fileNum, CString content)
 	CDragInformation::parseDragInfo(m_dragFileList, fileNum, content);
 	LOG((CLOG_DEBUG "drag info received, total drag file number: %i", m_dragFileList.size()));
 	
-	for (int i = 0; i < m_dragFileList.size(); ++i) {
+	for (size_t i = 0; i < m_dragFileList.size(); ++i) {
 		LOG((CLOG_DEBUG "dragging file %i name: %s", i + 1, m_dragFileList.at(i).c_str()));
 	}
